Refazendo estrutura do projeto para receber atualizacao da UI
