+++ src/App.tsx
@@ -1,35 +1,21 @@
-import React, { useState } from "react";
-import { ThemeProvider } from "styled-components";
-import light from "./styles/themes/light";
-import dark from "./styles/themes/dark";
-import GlobalStyles from "./styles/GlobalStyles";
-import ProgressCircle from "./components/ProgressCircle";
-import Header from "./components/Header";
-import Introduction from "./components/Introduction";
-import Transition from "./components/Transition";
-import ContainerHabilities from "./components/ContainerHabilities";
-import ContainerProjects from "./components/ContainerProjects";
-import Footer from "./components/Footer";
+import React from "react";
+
+import { Route, HashRouter, Switch } from "react-router-dom";
+import V1 from "./pages/V1";
+import V2 from "./pages/V2";
 
 function App() {
-  const [theme, setTheme] = useState(light);
-
-  const toggleTheme = () => {
-    setTheme(theme.title === "light" ? dark : light);
-  };
   return (
-    <ThemeProvider theme={theme}>
-      <div style={{ overflowX: "hidden" }}>
-        <GlobalStyles />
-        <ProgressCircle />
-        <Header toggleTheme={toggleTheme} />
-        <Introduction />
-        <Transition />
-        <ContainerHabilities />
-        <ContainerProjects />
-        <Footer />
-      </div>
-    </ThemeProvider>
+    <div style={{ overflowX: "hidden" }}>
+      <HashRouter>
+        <Switch>
+          <Route exact path="/" component={V2} />
+        </Switch>
+        <Switch>
+          <Route path="/old" component={V1} />
+        </Switch>
+      </HashRouter>
+    </div>
   );
 }
