+++ src/pages/V2/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+// import { Container } from './styles';
+
+const V2: React.FC = () => {
+  return <div />;
+}
+
+export default V2;
